% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Car}
\alias{Car}
\title{Stated Preferences for Car Choice}
\format{
A dataframe containing :
\itemize{
\item choice: choice of a vehicle among 6 propositions,
\item college: college education?,
\item hsg2: size of household greater than 2?
\item coml5: commute lower than 5 miles a day?,
\item typez: body type, one of regcar (regular car), sportuv (sport utility vehicle), sportcar, stwagon (station wagon), truck, van, for each proposition z from 1 to 6,
\item fuelz: fuel for proposition z, one of gasoline, methanol, cng (compressed natural gas), electric.,
\item pricez: price of vehicle divided by the logarithm of income,
\item rangez: hundreds of miles vehicle can travel between refuelings/rechargings,
\item accz: acceleration, tens of seconds required to reach 30 mph from stop,
\item speedz: highest attainable speed in hundreds of mph,
\item pollutionz: tailpipe emissions as fraction of those for new gas vehicle,
\item sizez: 0 for a mini, 1 for a subcompact, 2 for a compact and 3 for a mid--size or large vehicle,
\item spacez: fraction of luggage space in comparable new gas vehicle,
\item costz: cost per mile of travel (tens of cents) : home recharging for electric vehicle, station refueling otherwise,
\item stationz: fraction of stations that can refuel/recharge vehicle.
}
}
\source{
\href{https://wileyonlinelibrary.com/journal/jae/}{Journal of Applied Econometrics data archive}.
}
\description{
a sample of 4654 individuals
}
\references{
\insertRef{MCFA:TRAI:00}{mlogit}
}
\keyword{datasets}
