% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cracker}
\alias{Cracker}
\title{Choice of Brand for Crakers}
\format{A dataframe containing :
\itemize{
\item id: individuals identifiers,
\item choice: one of sunshine, keebler, nabisco, private,
\item disp.z: is there a display for brand z?
\item feat.z: is there a newspaper feature advertisement for brand z?
\item price.z: price of brand z.
}}
\source{
\href{https://www.amstat.org}{Journal of Business Economics and Statistics web site}.
}
\description{
a sample of 3292 individualscross-section
}
\references{
\insertRef{DIPA:NAUF:CHIN:94}{mlogit}

\insertRef{PAAP:FRAN:00}{mlogit}
}
\keyword{datasets}
