\name{Examples}
\docType{data}
\alias{Examples}
\alias{Elec.mxl}
\alias{Elec.mxl2}
\alias{Elec.mxl3}
\alias{Elec.mxl4}
\alias{Elec.mxl5}
\alias{Train.ml}
\alias{Train.mxlc}
\alias{Train.mxlu}
\alias{Fish.mprobit}
\alias{p1}
\alias{p2}
\title{Examples of mixed logit and multinomial probit models}
\description{

  This file contains several fitted mixed logit and multinomial probit
  models. These examples are used in the vignettes and in the man page
  and are stored in this file because they are pretty long to compute

}
\usage{data(Examples)}

\format{ These examples are obtained using the \code{"Electricity"} and
  the \code{"Train"}, \code{"Mode"} and \code{"Fishing"} data set.  }

\keyword{datasets}