% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Game}
\alias{Game}
\alias{Game2}
\title{Ranked data for gaming platforms}
\format{
A dataframe containing :

- ch.Platform: where `platform` is one of `Xbox`,
`PlayStation`, `PSPortable`, `GameCube`,
`GameBoy` and `PC`. This variables contain the ranking of
the platforms from 1 to 6,
- own.Platform: these 6 variables are dummies which indicate
whether the given plaform is already owned by the respondent,
- age: the age of the respondent,
- hours: hours per week spent on gaming.,
}
\source{
[Journal of Applied Econometrics data archive](https://wileyonlinelibrary.com/journal/jae/).
}
\description{
A sample of 91 Dutch individuals
}
\details{
The data are also provided in long format (use in this case
`data(Game2)`. In this case, the alternative and the choice
situation are respectively indicated in the `platform` and
`chid` variables.
}
\references{
\insertRef{FOK:PAAP:VAND:12}{mlogit}
}
\keyword{datasets}
