% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.data.R
\name{has.intercept}
\alias{has.intercept}
\alias{has.intercept.default}
\alias{has.intercept.formula}
\alias{has.intercept.Formula}
\alias{has.intercept.mlogit}
\title{Indicates whether the formula contains an intercept}
\usage{
has.intercept(object, ...)

\method{has.intercept}{default}(object, ...)

\method{has.intercept}{formula}(object, ...)

\method{has.intercept}{Formula}(object, rhs = NULL, ...)

\method{has.intercept}{mlogit}(object, ...)
}
\arguments{
\item{object}{the object}

\item{...}{further arguments}

\item{rhs}{for the Formula method the rhs for which one wants to
know if there is an intercept may be specified}
}
\description{
This is a generic which provide convenient methods for
formula/Formula object and for specific fitted models
}
\author{
Yves Croissant
}
\keyword{attribute}
