% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HC}
\alias{HC}
\title{Heating and Cooling System Choice in Newly Built Houses in California}
\format{
A dataframe containing :

- depvar: heating system, one of `gcc` (gas central heat with
cooling), `ecc` (electric central resistence heat with cooling), `erc`
(electric room resistence heat with cooling), `hpc` (electric heat
pump which provides cooling also), `gc` (gas central heat without
cooling), `ec` (electric central resistence heat without cooling), `er`
(electric room resistence heat without cooling),
- ich.z: installation cost of the heating portion of the
system,
- icca: installation cost for cooling,
- och.z: operating cost for the heating portion of the system,
- occa: operating cost for cooling,
- income: annual income of the household.
}
\source{
[Kenneth Train's home page](https://elsa.berkeley.edu/~train/).
}
\description{
A sample of 250 Californian households
}
\keyword{datasets}
