% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Heating}
\alias{Heating}
\title{Heating System Choice in California Houses}
\format{
A dataframe containing:

- idcase: id,
- depvar: heating system, one of gc (gas central), gr (gas room), ec
(electric central), er (electric room), hp (heat pump),
- ic.z: installation cost for heating system z (defined for the
5 heating systems),
- oc.z: annual operating cost for heating system z (defined for
the 5 heating systems),
- pb.z: ratio oc.z/ic.z ,
- income: annual income of the household,
- agehed: age of the household head
- rooms: numbers of rooms in the house,
}
\source{
[Kenneth Train's home page](https://elsa.berkeley.edu/~train/).
}
\description{
A sample of 900 Californian households#'
}
\keyword{datasets}
