% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{JapaneseFDI}
\alias{JapaneseFDI}
\title{Japanese Foreign Direct Investment in European Regions}
\format{
A dataframe containing :

- firm: the investment id,
- country: the country,
- region: the region (nuts1 nomenclature),
- choice: a dummy indicating the chosen region ,
- choice.c: the chosen country,
- wage: wage rate in the region,
- unemp: unemployment rate in the region,
- elig: is the country eligible to european subsidies,
- area: the area of the region,
- scrate:  social charge rate (country level),
- ctaxrate: corporate tax rate (country level),
- gdp: regional gdp,
- harris: harris' market potential,
- krugman: krugman's market potential,
- domind: domestic industry count,
- japind: japan industry count,
- network: network count.
}
\source{
kindly provided by Thierry Mayer
}
\description{
A sample of 452 Japanese production units in Europe
#'
}
\references{
\insertRef{HEAD:MAYE:04}{mlogit}
}
\keyword{datasets}
