\name{Ketchup}
\docType{data}
\alias{Ketchup}
\title{Choice of Brand for Ketchup }
\description{
  a cross-section

 \emph{number of observations} :  4956

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Ketchup)}
\format{A dataframe containing :
  \describe{
    \item{hid}{individuals identifiers}
    \item{id}{purchase identifiers}
    \item{choice}{one of heinz, hunts, delmonte, stb (store brand)}
    \item{price.z}{price of brand z}
  }
}
\source{
    Kim, Byong--Do,  Robert C.   Blattberg and  Peter E.  Rossi (1995) \dQuote{Modeling the distribution of price sensitivity and implications for optimal retail pricing}, \emph{Journal of Business Economics and Statistics}, \bold{13(3)}, 291.

}
\references{
      Journal of Business Economics and Statistics web site : \url{http://www.amstat.org/publications/jbes/}.
}

\keyword{datasets}
