% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/predict_slopes.R
\name{residuals.mlogit}
\alias{residuals.mlogit}
\alias{terms.mlogit}
\alias{model.matrix.mlogit}
\alias{update.mlogit}
\alias{print.mlogit}
\alias{summary.mlogit}
\alias{print.summary.mlogit}
\alias{idx.mlogit}
\alias{idx_name.mlogit}
\alias{fitted.mlogit}
\alias{coef.summary.mlogit}
\alias{predict.mlogit}
\alias{miscmethods.mlogit}
\alias{df.residual.mlogit}
\alias{coef.mlogit}
\title{Methods for mlogit objects}
\usage{
\method{residuals}{mlogit}(object, outcome = TRUE, ...)

\method{terms}{mlogit}(x, ...)

\method{model.matrix}{mlogit}(object, ...)

\method{update}{mlogit}(object, new, ...)

\method{print}{mlogit}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)

\method{summary}{mlogit}(object, ..., type = c("chol", "cov", "cor"))

\method{print}{summary.mlogit}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)

\method{idx}{mlogit}(x, n = NULL, m = NULL)

\method{idx_name}{mlogit}(x, n = NULL, m = NULL)

\method{fitted}{mlogit}(
  object,
  type = c("outcome", "probabilities", "linpred", "parameters"),
  outcome = NULL,
  ...
)

\method{coef}{summary.mlogit}(object, ...)

\method{predict}{mlogit}(object, newdata = NULL, returnData = FALSE, ...)
}
\arguments{
\item{outcome}{a boolean which indicates, for the \code{fitted} and the
\code{residuals} methods whether a matrix (for each choice, one
value for each alternative) or a vector (for each choice, only
a value for the alternative chosen) should be returned,}

\item{...}{further arguments.}

\item{x, object}{an object of class \code{mlogit}}

\item{new}{an updated formula for the \code{update} method,}

\item{digits}{the number of digits,}

\item{width}{the width of the printing,}

\item{type}{one of \code{outcome} (probability of the chosen
alternative), \code{probabilities} (probabilities for all the
alternatives), \code{parameters} for individual-level random
parameters for the fitted method, how the correlated random
parameters should be displayed : \code{"chol"} for the estimated
parameters (the elements of the Cholesky decomposition matrix),
\code{"cov"} for the covariance matrix and \code{"cor"} for the
correlation matrix and the standard deviations,}

\item{n, m}{see \code{\link[dfidx:idx]{dfidx::idx()}}}

\item{newdata}{a \code{data.frame} for the \code{predict} method,}

\item{returnData}{for the \code{predict} method, if \code{TRUE}, the data is
returned as an attribute,}
}
\description{
Miscellaneous methods for \code{mlogit} objects.
}
