\name{MixedExamples}
\docType{data}
\alias{MixedExamples}
\title{Examples of mixed logit model}
\description{

  This file contains several fitted mixed logit models. These examples
  are used in the vignettes and in the man page and are stored in this
  file because there are pretty long to compute

}
\usage{data(MixedExamples)}

\format{
  These examples are obtained using the \code{"Electricity"}
  and the \code{"Train"} data set.
}

\keyword{datasets}