% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{package}
\name{mlogit-package}
\alias{mlogit-package}
\title{mlogit package: estimation of random utility discrete choice models
by maximum likelihood}
\description{
mlogit provides a model description interface (enhanced
formula-data), a very versatile estimation function and a testing
infrastructure to deal with random utility models.
}
\details{
For a gentle and comprehensive introduction to the
    package, see the package's vignettes.

\insertRef{CROI:20}{mlogit}

\insertRef{TRAI:09}{mlogit}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=mlogit}
  \item \url{https://r-forge.r-project.org/projects/mlogit/}
}

}
\author{
\strong{Maintainer}: Yves Croissant \email{yves.croissant@univ-reunion.fr}

}
