% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Mode}
\alias{Mode}
\title{Mode Choice}
\format{
A dataframe containing :

- choice: one of car, carpool, bus or rail,
- cost.z: cost of mode z,
- time.z: time of mode z.
}
\source{
[Kenneth Train's home page](https://elsa.berkeley.edu/~train/).
}
\description{
A sample of 453 individuals for 4 transport modes.
}
\keyword{datasets}
