% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.data.R
\name{model.matrix.dfidx_mlogit}
\alias{model.matrix.dfidx_mlogit}
\title{Compute the model matrix for RUM}
\usage{
\method{model.matrix}{dfidx_mlogit}(object, ..., lhs = NULL, rhs = 1, dot = "separate")
}
\arguments{
\item{object}{the object}

\item{..., lhs, rhs, dot}{see the `Formula` method}
}
\description{
specific stuff compared to the model.matrix.dfidx method which
simply applies the Formula method
}
\author{
Yves Croissant
}
\keyword{attribute}
