% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOx}
\alias{NOx}
\title{Technologies to reduce NOx emissions}
\format{
A dataframe containing:

- chid: the plant id,
- alt: the alternative,
- id: the owner id,
- choice: the
chosen alternative,
- available: a dummy indicating that the alternative is
available,
- env: the regulatory environment, one of `'regulated'`,
`'deregulated'` and `'public'`,
- post: dummy for post-combustion polution control technology,
- cm: dummy for combustion modification technology,
- lnb: dummy for low NOx burners technology,
- age: age of the plant (in deviation from the mean age).,
- vcost: variable cost,
- kcost: capital cost.
}
\source{
[American Economic Association data archive](https://www.aeaweb.org/aer/).
}
\description{
A sample of 632 American production units
}
\references{
\insertRef{FOWL:10}{mlogit}
}
\keyword{datasets}
