% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_slopes.R
\name{preds}
\alias{preds}
\alias{summary.preds}
\title{Compute the predictions and their standard deviations}
\usage{
preds(object)

\method{summary}{preds}(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{further arguments for the summary method}
}
\value{
a \code{preds} object, which inherits the \code{dfidx} class and
therefore contains an index column.
}
\description{
\code{preds} is an alternative to the \code{predict} method
}
\details{
the standard errors are computed using the delta method.
}
\seealso{
\code{\link[=mlogit]{mlogit()}} for the estimation of a multinomial logit
model.
}
\author{
Yves Croissant
}
\keyword{regression}
