% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlogit.R, R/test.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dfidx}
\alias{gaze}
\alias{idx}
\alias{idx_name}
\alias{scoretest}
\alias{waldtest}
\alias{lrtest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dfidx}{\code{\link[dfidx]{dfidx}}, \code{\link[dfidx]{idx}}, \code{\link[dfidx]{idx_name}}}

  \item{lmtest}{\code{\link[lmtest]{lrtest}}, \code{\link[lmtest]{waldtest}}}

  \item{micsr}{\code{\link[micsr]{gaze}}, \code{\link[micsr]{scoretest}}}
}}

