% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RiskyTransport}
\alias{RiskyTransport}
\title{Risky Transportation Choices}
\format{
A dataframe containing:

- id: individual id,
- choice: 1 for the chosen mode,
- mode: one of `Helicopter`,`WaterTaxi`, `Ferry,
and `Hovercraft`,
- cost: the generalised cost of the transport mode,
- risk: the fatality rate, numbers of death per 100,000 trips,
- weight: weights,
- seats: ,
- noise: ,
- crowdness: ,
- convloc: ,
- clientele: ,
- chid: choice situation id,
- african: `yes` if born in Africa, `no` otherwise,
- lifeExp: declared life expectancy,
- dwage: declared hourly wage,
- iwage: imputed hourly wage,
- educ: level of education, one of `low` and `high`,
- fatalism: self-ranking of the degree of fatalism,
- gender: gender, one of `female` and `male`,
- age: age,
- haveChildren: `yes` if the traveler has children,
`no` otherwise,
- swim: `yes` if the traveler knows how to swim, `no,
otherwise.
}
\source{
[American Economic Association data archive](https://www.aeaweb.org/aer/).
}
\description{
1793 choices by 561 individuals of a transport mode at Freetwon
airport
}
\references{
\insertRef{LEON:MIGU:17}{mlogit}
}
\keyword{datasets}
