% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_slopes.R
\name{slps}
\alias{slps}
\alias{summary.slps}
\title{Compute the slopes and their standard deviations}
\usage{
slps(object)

\method{summary}{slps}(object, ...)
}
\arguments{
\item{object}{a fitted model}

\item{...}{further arguments for the summary method}
}
\value{
a \code{slps} object, which inherits the \code{tbl_df} \code{tbl} and
\code{data.frame} classes
}
\description{
\code{preds} is an alternative to the \code{predict} method
}
\details{
the standard errors are computed using the delta method.
}
\seealso{
\code{\link[=mlogit]{mlogit()}} for the estimation of a multinomial logit
model.
}
\author{
Yves Croissant
}
\keyword{regression}
