\name{Telephone}
\docType{data}
\alias{Telephone}
\title{Choice among residential telephone service options for local calling }
\description{
  a cross-section from 1984

 \emph{number of observations} :  434

 \emph{observation} :  households

   \emph{country} :  United-States

}
\usage{data(Train)}
\format{A dataframe containing :
  \describe{
    \item{choice}{a logical which indicates if the alternative has been chosen}
    \item{service}{a telephone service option}
    \item{household}{the household index}
    \item{cost}{the logarithm of the cost}
  }
}
\source{
  
  Walker J.L., Ben-Akiva M. and D. Bolduc (2007) \dQuote{Identification
    of parameters in normal error component logit-mixture (NECLM)
    models}, \emph{Journal of Applied Econometrics}, \bold{22},
    1095--1125.

  Train K.E., Mc Fadden D. and M. Ben-Akiva (1987) \dQuote{The demand
  for local telephone service: a fully discrete model of residential
  calling patterns and service choices}, \emph{Rand Journal of
  Economics}, \bold{18(1)}, 109--123
  
}
\references{
      Journal of Applied Econometrics data archive : \url{http://jae.wiley.com/jae/}.
}

\keyword{datasets}
