% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Train}
\alias{Train}
\title{Stated Preferences for Train Traveling}
\format{
A dataframe containing:

- id: individual identifient,
- choiceid: choice identifient,
- choice: one of 'A' or 'B',
- price_z: price of proposition z (z = 'A', 'B') in cents of
guilders,
- time_z: travel time of proposition z (z = 'A', 'B') in
minutes,
- comfort_z: comfort of proposition z (z = 'A', 'B'), 0, 1 or
2 in decreasing comfort order,
- change_z: number of changes for proposition z (z = 'A', 'B').
}
\source{
[Journal of Applied Econometrics data archive](https://wileyonlinelibrary.com/journal/jae/).
}
\description{
A sample of 235 Dutch individuals facing 2929 choice situations
}
\references{
\insertRef{BENA:BOLD:BRAD:93}{mlogit}

\insertRef{MEIJ:ROUW:06}{mlogit}
}
\keyword{datasets}
