\name{Tuna}
\docType{data}
\alias{Tuna}
\title{Choice of Brand for Tuna }
\description{
  a cross-section

 \emph{number of observations} :  13705

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Tuna)}
\format{A dataframe containing :
  \describe{
    \item{hid}{individuals identifiers}
    \item{id}{purchase identifiers}
    \item{choice}{one of skw (Starkist water), cosw (Chicken of the sea water), pw (store--specific private label water), sko (Starkist oil), coso (Chicken of the sea oil)}
    \item{price.z}{price of brand z}
  }
}
\source{
    Kim, Byong--Do,  Robert C.   Blattberg and  Peter E.  Rossi (1995) \dQuote{Modeling the distribution of price sensitivity and implications for optimal retail pricing}, \emph{Journal of Business Economics and Statistics}, \bold{13(3)}, 291.

}
\references{
      Journal of Business Economics and Statistics web site : \url{http://www.amstat.org/publications/jbes/}.
}

\keyword{datasets}
