% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Catsup}
\alias{Catsup}
\title{Choice of Brand for Catsup}
\format{A dataframe containing :
\itemize{
\item id: individuals identifiers,
\item choice: one of heinz41, heinz32, heinz28, hunts32,
\item disp.z: is there a display for brand z ?
\item feat.z: is there a newspaper feature advertisement for brand z?
\item price.z: price of brand z.
}}
\source{
\href{https://www.amstat.org}{Journal of Business Economics and Statistics web site}.
}
\description{
a sample of 2798 individuals
}
\references{
\insertRef{DIPA:NAUF:CHIN:94}{mlogit}
}
\keyword{datasets}
