\name{Catsup}
\docType{data}
\alias{Catsup}
\title{Choice of Brand for Catsup }
\description{
  a cross-section

 \emph{number of observations} :  2798

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Catsup)}
\format{A dataframe containing :
  \describe{
    \item{id}{individuals identifiers}
    \item{choice}{one of heinz41, heinz32, heinz28, hunts32}
    \item{disp.z}{is there a display for brand z ?}
    \item{feat.z}{is there a newspaper feature advertisement for brand z ?}
    \item{price.z}{price of brand z}
  }
}
\source{
    Jain, Dipak C.,  Naufel J.    Vilcassim and  Pradeep K.  Chintagunta (1994) \dQuote{A random--coefficients logit brand--choice model applied to panel data}, \emph{Journal of Business and Economics Statistics}, \bold{12(3)}, 317.

}
\references{
      Journal of Business Economics and Statistics web site : \url{http://www.amstat.org/publications/jbes/}.
}

\keyword{datasets}
