% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cracker}
\alias{Cracker}
\title{Choice of Brand for Crakers}
\format{
A dataframe containing :

- id: individuals identifiers,
- choice: one of sunshine, keebler, nabisco, private,
- disp.z: is there a display for brand z?
- feat.z: is there a newspaper feature advertisement for brand z?
- price.z: price of brand z.
}
\source{
[Journal of Business Economics and Statistics web site](https://www.amstat.org).
}
\description{
a sample of 3292 individualscross-section
}
\references{
\insertRef{DIPA:NAUF:CHIN:94}{mlogit}

\insertRef{PAAP:FRAN:00}{mlogit}
}
\keyword{datasets}
