% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Fishing}
\alias{Fishing}
\title{Choice of Fishing Mode}
\format{A dataframe containing :
\itemize{
\item mode: recreation mode choice, one of : beach, pier, boat and
charter,
\item price.beach: price for beach mode
\item price.pier: price for pier mode,
\item price.boat: price for private boat mode,
\item price.charter: price for charter boat mode,
\item catch.beach: catch rate for beach mode,
\item catch.pier: catch rate for pier mode,
\item catch.boat: catch rate for private boat mode,
\item catch.charter: catch rate for charter boat mode,
\item income: monthly income,
}}
\source{
\insertRef{CAME:TRIV:05}{mlogit}
}
\description{
A sample of 1182 individuals in the United-States for the choice of
4 alternative fishing modes.
}
\references{
\insertRef{HERR:KLIN:99}{mlogit}
}
\keyword{datasets}
