\name{Fishing}
\docType{data}
\alias{Fishing}
\title{Choice of Fishing Mode }
\description{
  a cross-section

 \emph{number of observations} :  1182

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Fishing)}
\format{A dataframe containing :
  \describe{
    \item{mode}{ recreation mode choice, on of : beach, pier, boat and charter}
    \item{price}{price for chosen alternative}
    \item{catch}{catch rate for chosen alternative}
    \item{pbeach}{price for beach mode}
    \item{ppier}{price for pier mode}
    \item{pboat}{price for private boat mode}
    \item{pcharter}{price for charter boat mode}
    \item{cbeach}{catch rate for beach mode}
    \item{cpier}{catch rate for pier mode}
    \item{cboat}{catch rate for private boat mode}
    \item{ccharter}{catch rate for charter boat mode}
    \item{income}{monthly income}
  }
}
\source{
     Herriges, J. A.  and  C. L.   Kling (1999) \dQuote{Nonlinear Income Effects in Random Utility Models}, \emph{Review of Economics and Statistics}, \bold{81}, 62-72.

}
\references{
    Cameron, A.C.  and  P.K.  Trivedi (2005) \emph{Microeconometrics : methods and applications}, Cambridge, pp. 463--466, 486 and 491--495.

}

\keyword{datasets}
