\name{Fishing}
\docType{data}
\alias{Fishing}
\title{Choice of Fishing Mode }
\description{
  a cross-section

 \emph{number of observations} :  1182

 \emph{observation} :  individuals

   \emph{country} :  United States

}
\usage{data(Fishing)}
\format{A dataframe containing :
  \describe{
    \item{mode}{recreation mode choice, one of : beach, pier, boat and charter}
    \item{price.beach}{price for beach mode}
    \item{price.pier}{price for pier mode}
    \item{price.boat}{price for private boat mode}
    \item{price.charter}{price for charter boat mode}
    \item{catch.beach}{catch rate for beach mode}
    \item{catch.pier}{catch rate for pier mode}
    \item{catch.boat}{catch rate for private boat mode}
    \item{catch.charter}{catch rate for charter boat mode}
    \item{income}{monthly income}
  }
}
\source{
     Herriges, J. A.  and C. L.  Kling (1999) \dQuote{Nonlinear Income
     Effects in Random Utility Models}, \emph{Review of Economics and
     Statistics}, \bold{81}, 62-72.
}
\references{
    Cameron, A.C.  and P.K.  Trivedi (2005) \emph{Microeconometrics :
    methods and applications}, Cambridge, pp. 463--466, 486 and
    491--495.
}

\keyword{datasets}
