\name{HC}
\docType{data}
\alias{HC}
\title{Heating and Cooling System Choice in Newly Built Houses in California }
\description{
  a cross-section

 \emph{number of observations} :  250

 \emph{observation} :  households

   \emph{country} :  California

}
\usage{data(HC)}
\format{A dataframe containing :
  \describe{
    \item{depvar}{heating system, one of gcc (gas central heat with cooling), ecc (electric central resistence heat with cooling), erc (electric room resistence heat with cooling), hpc (electric heat pump which provides cooling also), gc (gas central heat without cooling, ec (electric central resistence heat without cooling), er (electric room resistence heat without cooling)}
    \item{ich.z}{installation cost of the heating portion of the system}
    \item{icca}{installation cost for cooling}
    \item{och.z}{operating cost for the heating portion of the system}
    \item{occa}{operating cost for cooling}
    \item{income}{annual income of the household}
  }
}
\references{
      Kenneth Train's home page :  \url{http://elsa.berkeley.edu/~train/}.
}

\keyword{datasets}
