% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HC}
\alias{HC}
\title{Heating and Cooling System Choice in Newly Built Houses in California}
\format{A dataframe containing :
\itemize{
\item depvar: heating system, one of \code{gcc} (gas central heat with
cooling), \code{ecc} (electric central resistence heat with cooling), \code{erc}
(electric room resistence heat with cooling), \code{hpc} (electric heat
pump which provides cooling also), \code{gc} (gas central heat without
cooling), \code{ec} (electric central resistence heat without cooling), \code{er}
(electric room resistence heat without cooling),
\item ich.z: installation cost of the heating portion of the
system,
\item icca: installation cost for cooling,
\item och.z: operating cost for the heating portion of the system,
\item occa: operating cost for cooling,
\item income: annual income of the household.
}}
\source{
\href{http://elsa.berkeley.edu/~train/}{Kenneth Train's home page}.
}
\description{
A sample of 250 Californian households
}
\keyword{datasets}
