% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Heating}
\alias{Heating}
\title{Heating System Choice in California Houses}
\format{A dataframe containing:
\itemize{
\item idcase: id,
\item depvar: heating system, one of gc (gas central), gr (gas room), ec
(electric central), er (electric room), hp (heat pump),
\item ic.z: installation cost for heating system z (defined for the
5 heating systems),
\item oc.z: annual operating cost for heating system z (defined for
the 5 heating systems),
\item pb.z: ratio oc.z/ic.z ,
\item income: annual income of the household,
\item agehed: age of the household head
\item rooms: numbers of rooms in the house,
}}
\source{
\href{http://elsa.berkeley.edu/~train/}{Kenneth Train's home page}.
}
\description{
A sample of 900 Californian households#'
}
\keyword{datasets}
