% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{JapaneseFDI}
\alias{JapaneseFDI}
\title{Japanese Foreign Direct Investment in European Regions}
\format{A dataframe containing :
\itemize{
\item firm: the investment id,
\item country: the country,
\item region: the region (nuts1 nomenclature),
\item choice: a dummy indicating the chosen region ,
\item choice.c: the chosen country,
\item wage: wage rate in the region,
\item unemp: unemployment rate in the region,
\item elig: is the country eligible to european subsidies,
\item area: the area of the region,
\item scrate:  social charge rate (country level),
\item ctaxrate: corporate tax rate (country level),
\item gdp: regional gdp,
\item harris: harris' market potential,
\item krugman: krugman's market potential,
\item domind: domestic industry count,
\item japind: japan industry count,
\item network: network count.
}}
\source{
kindly provided by Thierry Mayer
}
\description{
A sample of 452 Japanese production units in Europe
#'
}
\references{
\insertRef{HEAD:MAYE:04}{mlogit}
}
\keyword{datasets}
