% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{JapaneseFDI}
\alias{JapaneseFDI}
\title{Japanese Foreign Direct Investment in European Regions}
\format{A dataframe containing : \describe{

\item{firm}{the investment id}
\item{country}{the country}
\item{region}{the region (nuts1 nomenclature)}
\item{choice}{a dummy indicating the chosen region }
\item{choice.c}{the chosen country}
\item{wage}{wage rate in the region}
\item{unemp}{unemployment rate in the region}
\item{elig}{is the country eligible to european subsidies}
\item{area}{the area of the region}
\item{scrate}{ social charge rate (country level)}
\item{ctaxrate}{corporate tax rate (country level)}
\item{gdp}{regional gdp}
\item{harris}{harris' market potential}
\item{krugman}{krugman's market potential}
\item{domind}{domestic industry count}
\item{japind}{japan industry count}
\item{network}{network count}
}}
\source{
.
}
\description{
a cross-section
}
\details{
\emph{number of observations} : 25764

\emph{country} : Europe
}
\references{
Head, Keith and Thierry Mayer (2004) \dQuote{Market potential
and the location of Japanese investment in the european union}, \emph{the
Review of Economics and Statistics}, \bold{86(4)}, 959-972.
}
\keyword{datasets}
