% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mlogit-deprecated}
\alias{mlogit-deprecated}
\alias{mlogit.data}
\alias{mFormula}
\alias{mFormula.formula}
\alias{mFormula.default}
\alias{model.matrix.mFormula}
\alias{is.mFormula}
\alias{index.dfidx}
\alias{index.mlogit}
\title{Some deprecated functions, especially `mlogit.data`, `index` and
`mFormula`}
\usage{
mlogit.data(
  data,
  choice = NULL,
  shape = c("long", "wide"),
  varying = NULL,
  sep = ".",
  alt.var = NULL,
  chid.var = NULL,
  alt.levels = NULL,
  id.var = NULL,
  group.var = NULL,
  opposite = NULL,
  drop.index = FALSE,
  ranked = FALSE,
  subset = NULL,
  ...
)

mFormula(object)

\method{mFormula}{formula}(object)

\method{mFormula}{default}(object)

\method{model.matrix}{mFormula}(object, data, ...)

is.mFormula(object)

\method{index}{dfidx}(x, ...)

\method{index}{mlogit}(x, ...)
}
\arguments{
\item{data}{a `data.frame`,}

\item{choice}{the variable indicating the choice made: it can be
either a logical vector, a numerical vector with 0 where the
alternative is not chosen, a factor with level 'yes' when the
alternative is chosen}

\item{shape}{the shape of the `data.frame`: whether `long` if each
row is an alternative or `wide` if each row is an observation,}

\item{varying}{the indexes of the variables that are alternative
specific,}

\item{sep}{the seperator of the variable name and the alternative
name (only relevant for a `wide` `data.frame`),}

\item{alt.var}{the name of the variable that contains the
alternative index (for a `long` `data.frame` only) or the name
under which the alternative index will be stored (the default
name is `alt`),}

\item{chid.var}{the name of the variable that contains the choice
index or the name under which the choice index will be stored,}

\item{alt.levels}{the name of the alternatives: if null, for a
`wide` data.frame, they are guessed from the variable names and
the choice variable (both should be the same), for a `long`
`data.frame`, they are guessed from the `alt.var` argument,}

\item{id.var}{the name of the variable that contains the individual
index if any,}

\item{group.var}{the name of the variable that contains the group
index if any,}

\item{opposite}{returns the opposite of the specified variables,}

\item{drop.index}{should the index variables be dropped from the
`data.frame`,}

\item{ranked}{a logical value which is true if the response is a
rank,}

\item{subset}{a logical expression which defines the subset of
observations to be selected,}

\item{...}{further arguments passed to `reshape`.}

\item{x, object}{a `formula`, a `dfidx` or a `mlogit` object,}

\item{drop}{a boolean, equal to `FALSE` if one wants that a
`data.frame` is always returned,}
}
\value{
`mlogit.data` now returns a `dfidx` object, `mFormula`
    simply calls [Formula::Formula()] and returns a `Formula`
    object.
}
\description{
`mlogit.data` is deprecated, use [dfidx::dfidx()] instead,
`mFormula` is replaced by [Formula::Formula()] and [zoo::index()]
by `idx`.
}
\seealso{
[stats::reshape()]
}
\author{
Yves Croissant
}
\keyword{attribute}
