% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Mode}
\alias{Mode}
\title{Mode Choice}
\format{A dataframe containing :
\itemize{
\item choice: one of car, carpool, bus or rail,
\item cost.z: cost of mode z,
\item time.z: time of mode z.
}}
\source{
\href{http://elsa.berkeley.edu/~train/}{Kenneth Train's home page}.
}
\description{
A sample of 453 individuals for 4 transport modes.
}
\keyword{datasets}
