\name{Mode}
\docType{data}
\alias{Mode}
\title{Mode Choice}
\description{
  a cross-section

 \emph{number of observations} :  453

 \emph{observation} :  individuals

}
\usage{data(Mode)}
\format{A dataframe containing :
  \describe{
    \item{choice}{one of car, carpool, bus or rail}
    \item{cost.z}{cost of mode z}
    \item{time.z}{time of mode z}
  }
}
\references{
      Kenneth Train's home page :  \url{http://elsa.berkeley.edu/~train/}.
}

\keyword{datasets}
