\name{ModeCanada}
\docType{data}
\alias{ModeCanada}
\title{Mode Choice for the Montreal-Toronto Corridor}
\description{
  a cross-section

 \emph{number of observations} :  3880

 \emph{observation} :  individuals

}
\usage{data(ModeCanada)}
\format{A dataframe containing :
  \describe{
    \item{case}{the individual index}
    \item{alt}{the alternative, one of train, car, bus and air,}
    \item{choice}{one if the mode is chosen, zero otherwise,}
    \item{cost}{monetary cost,}
    \item{ivt}{in vehicule time,}
    \item{ovt}{out vehicule time,}
    \item{frequency}{frequency,}
    \item{income}{income,}
    \item{urban}{urban,}
    \item{noalt}{the number of alternatives available.}
  }
}
\references{

  Bhat, Chandra R. (1995) \dQuote{A heteroscedastic extreme value model
  of intercity travel mode choice}, \emph{Transportation Research Part
  B}, \bold{29(6)}, 471-483.
  
  Koppelman Franck S.  and Chieh-Hua Wen (2001) \dQuote{The paired
    combinatorial logit model:properties, estimation and application},
    \emph{Transportation Research Part B}, 75-89.
  
  Wen, Chieh-Hua and Franck S. Koppelman (2001) \dQuote{The generalized
    nested logit model}, \emph{Transportation Research Part B}, 627-641.
  

  
}

\keyword{datasets}
