% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NOx}
\alias{NOx}
\title{Technologies to reduce NOx emissions}
\format{A dataframe containing:
\itemize{
\item chid: the plant id,
\item alt: the alternative,
\item id: the owner id,
\item choice: the
chosen alternative,
\item available: a dummy indicating that the alternative is
available,
\item env: the regulatory environment, one of \code{'regulated'},
\code{'deregulated'} and \code{'public'},
\item post: dummy for post-combustion polution control technology,
\item cm: dummy for combustion modification technology,
\item lnb: dummy for low NOx burners technology,
\item age: age of the plant (in deviation from the mean age).,
\item vcost: variable cost,
\item kcost: capital cost.
}}
\source{
\href{http://aeaweb.org/aer/}{American Economic Association data archive}.
}
\description{
A sample of 632 American production units
}
\references{
\insertRef{FOWL:10}{mlogit}
}
\keyword{datasets}
