% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/test.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{index}
\alias{waldtest}
\alias{lrtest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lmtest}{\code{\link[lmtest]{lrtest}}, \code{\link[lmtest]{waldtest}}}

  \item{zoo}{\code{\link[zoo]{index}}}
}}

