% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RiskyTransport}
\alias{RiskyTransport}
\title{Risky Transportation Choices}
\format{A dataframe containing:
\itemize{
\item id: individual id,
\item choice: 1 for the chosen mode,
\item mode: one of \code{Helicopter},\code{WaterTaxi}, \verb{Ferry, and }Hovercraft`,
\item cost: the generalised cost of the transport mode,
\item risk: the fatality rate, numbers of death per 100,000 trips,
\item weight: weights,
\item seats: ,
\item noise: ,
\item crowdness: ,
\item convloc: ,
\item clientele: ,
\item chid: choice situation id,
\item african: \code{yes} if born in Africa, \code{no} otherwise,
\item lifeExp: declared life expectancy,
\item dwage: declared hourly wage,
\item iwage: imputed hourly wage,
\item educ: level of education, one of \code{low} and \code{high},
\item fatalism: self-ranking of the degree of fatalism,
\item gender: gender, one of \code{female} and \code{male},
\item age: age,
\item haveChildren: \code{yes} if the traveler has children,
\code{no} otherwise,
\item swim: \code{yes} if the traveler knows how to swim, `no,
otherwise.
}}
\source{
\href{http://aeaweb.org/aer/}{American Economic Association data archive}.
}
\description{
1793 choices by 561 individuals of a transport mode at Freetwon
airport
}
\references{
\insertRef{LEON:MIGU:17}{mlogit}
}
\keyword{datasets}
